Added filter for auto.master
Modified 411 transport to encode content of the file
and filter before encrypting them. This helps with the
problem of stripping whitespaces, and returns the content
of the files exactly as they should be.

Also modified the password and group filters just a little
to make them return the correct whitespaces