Another installment of "Why did you do that?"
Redhat decided to patch python 2.4.3 and to move some function which once where in
os.py as a built-in methods (this is a python 2.6 code) with no version change
though.

This obviously messes up when you run foundation-python insde anaconda which has
the pythonpath pointing to the redhat libraries first.

Here some more description of the bug:
https://bugzilla.redhat.com/show_bug.cgi?id=750555