Fix updates and base path
Updates and base path to retreive the RPM from the mirror did not include
the last trailing / so wget in recursive mode was fetching not just
/centos/6.3/updates/x86_64/Packages/
but also all the other /centos/6.3/updates/x86_64/drpms/ etc.
The drpms folder contains 10GB of stuff.
So it fills the disk during comilation.