removed useless pice of code from report host interface
This piece of code is useless since its functionalities are moved in the
rocks sync host vlan.

Vlan interfaces now are not managed anymore by the report command but they
are brought up dynamically when we start VM using rocks sync host vlan.

Jim Hayes reported that when you have a ipmi interface on a subnet ( e.g.
admin) and another interface tagged with a vlan on the same subnet (e.g.
admin) the rocks report host create an interface called ipmi.vlanid.

Signed-off-by: Luca Clementi <luca.clementi@gmail.com>