commands: report interface: when using dhcp do not put the mac address
When an interface is in DHCP mode we should not put the mac address
in the ifup configuration files. DHCP is mostly used in vritual machine
or in cloud system (e.g. EC2) so it is simpler if we don't put the
MAC address
(cherry picked from commit 4f944d186df0dc0bcd9186e423377f002ccc8a85)