commands: save the old value of the attribute when you change value
When the user change the value of an attribute called 'name' the
old value is saved in an attribute called 'name'_old
This is used by the reconfigure procedure.

This code is pretty bad, lots of code duplication. I didn't have
time to fix previous structure of command to avoid code duplication.